centralise logic
