+++ src/renderer/modals/Send/steps/StepConnectDevice.js
@@ -1,6 +1,5 @@
 // @flow
-import React, { useEffect, useState } from "react";
-import { getAccountCurrency } from "@ledgerhq/live-common/lib/account/helpers";
+import React from "react";
 import TrackPage from "~/renderer/analytics/TrackPage";
 import GenericStepConnectDevice from "./GenericStepConnectDevice";
 import type { StepProps } from "../types";
@@ -14,22 +13,8 @@ export default function StepConnectDevice({
   onOperationBroadcasted,
   onTransactionError,
   setSigned,
+  currencyName,
 }: StepProps) {
-  const [currencyName, setCurrencyName] = useState("");
-
-  useEffect(() => {
-    if (account) {
-      const currency = getAccountCurrency(account);
-
-      const currencyName = currency
-        ? currency.type === "TokenCurrency"
-          ? currency.parentCurrency.name
-          : currency.name
-        : undefined;
-
-      setCurrencyName(currencyName);
-    }
-  }, [account]);
   return (
     <>
       <TrackPage category="Send Flow" name="Step ConnectDevice" currencyName={currencyName} />
