+++ src/renderer/modals/Send/steps/StepRecipient.js
@@ -1,8 +1,7 @@
 // @flow
 
-import React, { PureComponent, useEffect, useState } from "react";
+import React, { PureComponent } from "react";
 import { getMainAccount } from "@ledgerhq/live-common/lib/account";
-import { getAccountCurrency } from "@ledgerhq/live-common/lib/account/helpers";
 
 import TrackPage from "~/renderer/analytics/TrackPage";
 import Box from "~/renderer/components/Box";
@@ -32,24 +31,10 @@ const StepRecipient = ({
   bridgePending,
   maybeRecipient,
   onResetMaybeRecipient,
+  currencyName,
 }: StepProps) => {
-  const [currencyName, setCurrencyName] = useState("");
-
-  useEffect(() => {
-    if (account) {
-      const currency = getAccountCurrency(account);
-
-      const currencyName = currency
-        ? currency.type === "TokenCurrency"
-          ? currency.parentCurrency.name
-          : currency.name
-        : undefined;
-
-      setCurrencyName(currencyName);
-    }
-  }, [account]);
-  const mainAccount = account ? getMainAccount(account, parentAccount) : null;
   if (!status) return null;
+  const mainAccount = account ? getMainAccount(account, parentAccount) : null;
 
   return (
     <Box flow={4}>
