+++ src/renderer/modals/Send/steps/StepSummary.js
@@ -57,7 +57,7 @@ const WARN_FROM_UTXO_COUNT = 50;
 
 export default class StepSummary extends PureComponent<StepProps> {
   render() {
-    const { account, parentAccount, transaction, status } = this.props;
+    const { account, parentAccount, transaction, status, currencyName } = this.props;
     if (!account) return null;
     const mainAccount = getMainAccount(account, parentAccount);
     if (!mainAccount || !transaction) return null;
@@ -71,12 +71,6 @@ export default class StepSummary extends PureComponent<StepProps> {
     const hasNonEmptySubAccounts =
       account.type === "Account" &&
       (account.subAccounts || []).some(subAccount => subAccount.balance.gt(0));
-
-    const currencyName = currency
-      ? currency.type === "TokenCurrency"
-        ? currency.parentCurrency.name
-        : currency.name
-      : undefined;
 
     // $FlowFixMe
     const memo = transaction.memo;
